refactor(mcp-gateway): compose style-agent tools through gateway
- McpGateway owns StyleAgentEndpoints and constructs SearchEngine
  internally from StyleAgentConfig — no injection of internals
- Unified /mcp endpoint with auth: hello + search_code tools
- build_app is generic over MCP service type (tower::Service bounds)
  so web crate doesn't depend on style-agent internals
- Removed `enabled` flag from StyleAgentConfig — search engine always
  initialises (empty db returns zero results)
- Removed `layer` and `uses` filters from SearchCodeParams
- Removed `review_code` tool (not ready)
- Decomposed style-agent-server into endpoints.rs, standalone_server.rs,
  config.rs modules
- StyleAgentServer wraps StyleAgentEndpoints for standalone use
