+++ charts/galoy-agents/templates/deployment.yaml
@@ -135,8 +135,6 @@ spec:
           value: {{ .Values.galoyAgents.githubAllowedTeams | join "," | quote }}
         {{- end }}
         {{- if .Values.galoyAgents.styleAgent.indexHash }}
-        - name: STYLE_AGENT_ENABLED
-          value: "true"
         - name: STYLE_AGENT_DB_PATH
           value: "/data/style-agent/style-agent.db"
         {{- end }}
