+++ cli/Cargo.toml
@@ -9,9 +9,8 @@ path = "src/main.rs"
 
 [dependencies]
 galoy-agents-domain = { path = "../domain" }
+galoy-agents-mcp-gateway = { path = "../mcp-gateway" }
 galoy-agents-web = { path = "../web" }
-style-agent-core = { workspace = true }
-style-agent-server = { workspace = true }
 anyhow = "1"
 axum = { workspace = true }
 clap = { version = "4", features = ["derive", "env"] }
