+++ cli/src/main.rs
@@ -1,7 +1,5 @@
 mod config;
 
-use std::sync::Arc;
-
 use clap::Parser;
 
 use config::{Config, EnvSecrets};
@@ -79,12 +77,10 @@ async fn main() -> anyhow::Result<()> {
         secure_cookies: config.server.secure_cookies,
     };
 
-    let mut router = galoy_agents_web::server::build_app(&server_config, &pool, app_state);
+    let mcp_service =
+        galoy_agents_mcp_gateway::McpGateway::service(app_state.app.clone(), &config.style_agent)?;
 
-    if config.style_agent.enabled {
-        let style_agent_router = init_style_agent(&config.style_agent)?;
-        router = router.nest("/style-agent", style_agent_router);
-    }
+    let router = galoy_agents_web::server::build_app(&server_config, &pool, app_state, mcp_service);
 
     let addr: std::net::SocketAddr =
         format!("{}:{}", config.server.host, config.server.port).parse()?;
@@ -94,26 +90,4 @@ async fn main() -> anyhow::Result<()> {
     axum::serve(listener, router).await?;
 
     Ok(())
-}
-
-/// Initialise the style-agent search engine and return its axum router.
-fn init_style_agent(config: &config::StyleAgentConfig) -> anyhow::Result<axum::Router> {
-    use std::path::Path;
-    use style_agent_core::embedder::Embedder;
-    use style_agent_core::search::SearchEngine;
-    use style_agent_core::store::VectorStore;
-
-    tracing::info!(db_path = %config.db_path, "Initialising style-agent");
-
-    let store = VectorStore::new(Path::new(&config.db_path))?;
-    store.ensure_collection()?;
-    store.ensure_anti_pattern_tables()?;
-
-    let embedder = Embedder::new()?;
-    let search_engine = Arc::new(SearchEngine::new(embedder, store));
-
-    let router = style_agent_server::router(search_engine);
-
-    tracing::info!("Style-agent mounted at /style-agent/mcp");
-    Ok(router)
 }
