+++ galoy-agents.yml
@@ -7,5 +7,4 @@ oauth:
   github_redirect_uri: "http://localhost:4200/auth/github/callback"
   github_allowed_teams: []
 style_agent:
-  enabled: false
   db_path: "/data/style-agent/style-agent.db"
