+++ mcp-gateway/Cargo.toml
@@ -7,6 +7,9 @@ edition.workspace = true
 
 [dependencies]
 galoy-agents-domain = { path = "../domain" }
+style-agent-server = { workspace = true }
+anyhow = { workspace = true }
+axum = { workspace = true }
 http = "1"
 rmcp = { version = "1.2", features = ["server", "transport-streamable-http-server", "macros"] }
 schemars = "0.8"
