+++ web/Cargo.toml
@@ -23,6 +23,7 @@ sqlx = { version = "0.8", default-features = false, features = ["runtime-tokio-r
 thiserror = "2.0"
 time = "0.3"
 tower-http = { version = "0.6", features = ["cors"] }
+tower-service = "0.3"
 tokio = { version = "1", features = ["net"] }
 tower-sessions = "0.15"
 tracing = "0.1"
