+++ web/src/server.rs
@@ -1,7 +1,9 @@
+use std::convert::Infallible;
+
 use tower_sessions::{cookie::SameSite, SessionManagerLayer};
 use tracing::instrument;
 
-use galoy_agents_mcp_gateway::McpGateway;
+use galoy_agents_mcp_gateway::{McpGateway, StyleAgentConfig};
 
 use crate::auth::session_store::PgSessionStore;
 use crate::AppState;
@@ -15,15 +17,27 @@ pub struct ServerConfig {
 /// Build the axum [`Router`] with all web routes, MCP gateway, auth, and
 /// session middleware applied.
 ///
-/// Returns a `Router<()>` that can be extended (e.g. with
-/// [`axum::Router::nest`]) before binding to a listener.
-pub fn build_app(config: &ServerConfig, pool: &sqlx::PgPool, app_state: AppState) -> axum::Router {
+/// The `mcp_service` is the pre-built MCP gateway service (from
+/// [`McpGateway::service`]) that will be mounted at `/mcp`.
+pub fn build_app<M>(
+    config: &ServerConfig,
+    pool: &sqlx::PgPool,
+    app_state: AppState,
+    mcp_service: M,
+) -> axum::Router
+where
+    M: tower_service::Service<axum::extract::Request, Error = Infallible>
+        + Clone
+        + Send
+        + Sync
+        + 'static,
+    M::Response: axum::response::IntoResponse,
+    M::Future: Send + 'static,
+{
     let session_store = PgSessionStore::new(pool);
     let session_layer = SessionManagerLayer::new(session_store)
         .with_same_site(SameSite::Lax)
         .with_secure(config.secure_cookies);
-
-    let mcp_service = McpGateway::service(app_state.app.clone());
 
     crate::router()
         .nest_service("/mcp", mcp_service)
@@ -39,7 +53,8 @@ pub async fn run(
     pool: &sqlx::PgPool,
     app_state: AppState,
 ) -> anyhow::Result<()> {
-    let app = build_app(&config, pool, app_state);
+    let mcp_service = McpGateway::service(app_state.app.clone(), &StyleAgentConfig::default())?;
+    let app = build_app(&config, pool, app_state, mcp_service);
 
     let addr: std::net::SocketAddr = format!("{}:{}", config.host, config.port)
         .parse()
